+++ server/src/graphql/user/type-defs.js
@@ -57,7 +57,7 @@ const typeDefs = `
     avatar: String
     city: String
     country: String
-    coordinates: [Int]
+    coordinates: [Float]
   }
 
   # QUERIES
